Added error checks to most installations and modifications

+++ 12.10/prepare_install_2_3.sh
@@ -81,7 +81,13 @@ function installDependencies()
     echo "-- Installing installation dependencies..."
     echo ""
 
-	sudo apt-get -y -qq install dialog software-properties-common > /dev/null 2>&1
+	sudo apt-get -y install dialog software-properties-common > /dev/null 2>&1
+	
+	if [ $? != 0 ];
+	then
+	    echo "Error installing dependencies. Installation aborted."
+	    exit
+	fi
 }
 
 function fixLocaleBug()
@@ -102,7 +108,12 @@ function fixLocaleBug()
 	echo "LC_MESSAGES=\"C\"" | sudo tee -a $ENVIRONMENT_FILE > /dev/null 2>&1
 	echo "LC_ALL=\"en_US.UTF-8\"" | sudo tee -a $ENVIRONMENT_FILE > /dev/null 2>&1
 	
-	showInfo "Locale environment bug fixed"
+	if [ $? == 0 ];
+	then
+	    showInfo "Ubuntu locale environment bug fixed"
+	else
+	    showError "ERROR: Ubuntu locale environment bug could not be fixed"
+	fi
 }
 
 function applyXbmcNiceLevelPermissions()
@@ -113,7 +124,13 @@ function applyXbmcNiceLevelPermissions()
 	fi
 
 	echo "$USER             -       nice            -1" | sudo tee -a $SYSTEM_LIMITS_FILE> /dev/null 2>&1
-	showInfo "Allowed XBMC to prioritize threads"
+	
+	if [ $? == 0 ];
+	then
+	    showInfo "Allowed XBMC to prioritize threads"
+	else
+	    showError "ERROR: Failed to allow XBMC to prioritize threads"
+	fi
 }
 
 function addUserToRequiredGroups()
@@ -129,7 +146,13 @@ function addXbmcPpa()
     showInfo "Adding Wsnipex xbmc-xvba PPA..."
     sudo mkdir -p $HOME_DIRECTORY".gnupg/" > /dev/null 2>&1
 	sudo add-apt-repository -y $XBMC_PPA > /dev/null 2>&1
-	showInfo "Wsnipex xbmc-xvba PPA added"
+	
+	if [ $? == 0 ];
+	then
+	    showInfo "Wsnipex xbmc-xvba PPA added"
+	else
+	    showError "ERROR: Could not add xbmc-xvba PPA"
+	fi
 }
 
 function distUpgrade()
@@ -137,7 +160,13 @@ function distUpgrade()
     showInfo "Updating Ubuntu with latest packages (may take a while)..."
 	sudo apt-get -qq update > /dev/null 2>&1
 	sudo apt-get -y -qq dist-upgrade > /dev/null 2>&1
-	showInfo "Ubuntu installation updated"
+	
+	if [ $? == 0 ];
+	then
+	    showInfo "Ubuntu installation updated"
+	else
+	    showError "ERROR: Ubuntu dist-upgrade failed"
+	fi
 }
 
 function installXinit()
@@ -148,7 +177,13 @@ function installXinit()
     if [ $? == 1 ];
     then
         sudo apt-get -y -qq install xinit > /dev/null 2>&1
-	    showInfo "Xinit installed"
+        
+        if [ $? == 0 ];
+	    then
+	        showInfo "Xinit installed"
+	    else
+	        showError "ERROR: Xinit could not be installed"
+	    fi
     else
         showInfo "Skipping. Xinit already installed"
     fi
@@ -161,15 +196,29 @@ function installPowerManagement()
     mkdir -p $TEMP_DIRECTORY > /dev/null
 	cd $TEMP_DIRECTORY
 	sudo apt-get -y -qq install policykit-1 upower udisks acpi-support > /dev/null 2>&1
+	
+	if [ $? != 0 ];
+    then
+        showError "ERROR: A problem occured while instlling power management packages"
+    else
+        showInfo "Power management packages installed"
+    fi
+	
 	wget -q $SCRIPTS_DIR_URL"custom-actions.pkla" > /dev/null 2>&1
 	sudo mkdir -p $POWERMANAGEMENT_DIR > /dev/null 2>&1
 	
 	if [ -f ./custom-actions.pkla ];
 	then
         sudo mv custom-actions.pkla $POWERMANAGEMENT_DIR > /dev/null 2>&1
-	    showInfo "Power management packages successfully installed"
+        
+        if [ $? == 0 ];
+	    then
+	        showInfo "XBMC power management features enabled"
+	    else
+	        showError "ERROR: XBMC power management features could not be enabled"
+	    fi
 	else
-	    showError "Could not enable XBMC power management features"  
+	    showError "ERROR: XBMC power management file could not be downloaded"
 	fi
 }
 
@@ -177,15 +226,27 @@ function installAudio()
 {
     showInfo "Installing audio packages....\n!! Please make sure no used channels are muted !!"
 	sudo apt-get -y -qq install linux-sound-base alsa-base alsa-utils libasound2 > /dev/null 2>&1
-    sudo alsamixer
-    showInfo "Audio packages successfully installed"
+	
+	if [ $? == 0 ];
+    then
+        showInfo "Audio packages successfully installed"
+        sudo alsamixer
+    else
+        showError "ERROR: Audio packages could not be installed"
+    fi
 }
 
 function installLirc()
 {
     showInfo "Installing lirc"
-	sudo apt-get -y install lirc
-    showInfo "Lirc successfully installed"
+	sudo apt-get -y -qq install lirc
+	
+	if [ $? == 0 ];
+    then
+        showInfo "Lirc successfully installed"
+    else
+        showError "ERROR: Lirc installation failed"
+    fi
 }
 
 function installTvHeadend()
@@ -193,14 +254,26 @@ function installTvHeadend()
     showInfo "Adding jabbors hts-stable PPA..."
     sudo mkdir -p $HOME_DIRECTORY".gnupg/" > /dev/null 2>&1
 	sudo add-apt-repository -y $HTS_TVHEADEND_PPA > /dev/null 2>&1
-	showInfo "Jabbors hts-stable PPA added"
 	
-    distUpgrade
+	if [ $? == 0 ];
+    then
+        showInfo "Jabbors hts-stable tvheadend PPA added"
+        
+        distUpgrade
     
-    showInfo "Installing hts tvheadend..."
-    sudo apt-get -y -qq install tvheadend > /dev/null 2>&1
-    sudo dpkg-reconfigure tvheadend
-    showInfo "Hts tvheadend installed"
+        showInfo "Installing hts tvheadend..."
+        sudo apt-get -y -qq install tvheadend > /dev/null 2>&1
+        sudo dpkg-reconfigure tvheadend
+        
+        if [ $? == 0 ];
+        then
+            showInfo "Hts tvheadend successfully installed"
+        else
+            showError "ERROR: Hts tvheadend could not be installed"
+        fi
+    else
+        showError "ERROR: Jabbors hts-stable tvheadend PPA could not be added"
+    fi
 }
 
 function installOscam()
@@ -208,13 +281,25 @@ function installOscam()
     showInfo "Adding oscam PPA..."
     sudo mkdir -p $HOME_DIRECTORY".gnupg/" > /dev/null 2>&1
 	sudo add-apt-repository -y $OSCAM_PPA > /dev/null 2>&1
-	showInfo "Oscam PPA added"
 	
+	if [ $? == 0 ];
+    then
+        showInfo "Oscam PPA successfully added"
+    else
+        showError "ERROR: Oscam PPA could not be added"
+    fi
+
     distUpgrade
     
     showInfo "Installing oscam..."
     sudo apt-get -y -qq install oscam > /dev/null 2>&1
-    showInfo "Oscam installed"
+    
+    if [ $? == 0 ];
+    then
+        showInfo "Oscam successfully installed"
+    else
+        showError "ERROR: Oscam could not be installed"
+    fi
 }
 
 function installXbmc()
@@ -225,7 +310,13 @@ function installXbmc()
     if [ $? == 1 ];
     then
 	    sudo apt-get -y -qq install xbmc > /dev/null 2>&1
-        showInfo "XBMC successfully installed"
+	    
+	    if [ $? == 0 ];
+        then
+            showInfo "XBMC successfully installed"
+        else
+            showError "ERROR: XBMC could not be installed"
+        fi
     else
         showInfo "Skipping. XBMC already installed"
     fi
@@ -253,9 +344,15 @@ function enableDirtyRegionRendering()
 	if [ -f ./dirty_region_rendering.xml ];
 	then
         mv dirty_region_rendering.xml $XBMC_ADVANCEDSETTINGS_FILE > /dev/null 2>&1
-        showInfo "XBMC dirty-region-rendering enabled"
+        
+        if [ $? == 0 ];
+        then
+            showInfo "XBMC dirty region rendering enabled"
+        else
+            showError "ERROR: XBMC dirty region rendering could not be enabled"
+        fi
     else
-        showError "XBMC dirty-region-rendering could not be enabled"
+        showError "XBMC dirty region rendering file could not be downloaded"
     fi
 }
 
@@ -274,9 +371,15 @@ function installXbmcAddonRepositoriesInstaller()
     if [ -f ./plugin.program.repo.installer-1.0.5.tar.gz ];
     then
         tar -xvzf ./plugin.program.repo.installer-1.0.5.tar.gz -C $XBMC_ADDONS_DIR > /dev/null 2>&1
-	    showInfo "Addon Repositories Installer addon successfully installed"
+        
+        if [ $? == 0 ];
+        then
+            showInfo "Addon Repositories Installer addon successfully installed"
+        else
+            showError "ERROR: Addon Repositories Installer addon could not be installed"
+        fi
     else
-	    showError "Addon Repositories Installer addon could not be installed"
+	    showError "Addon Repositories Installer addon could not be downloaded"
     fi
 }
 
@@ -288,26 +391,37 @@ function installVideoDriver()
     if [ $? == 1 ];
     then
         sudo apt-get -y install $VIDEO_DRIVER > /dev/null 2>&1
-
-        if [ $VIDEO_MANUFACTURER == "ati" ];
-        then
-	        sudo aticonfig --initial -f > /dev/null 2>&1
-	        sudo aticonfig --sync-vsync=on > /dev/null 2>&1
-	        sudo aticonfig --set-pcs-u32=MCIL,HWUVD_H264Level51Support,1 > /dev/null 2>&1
-
-	        dialog --title "Disable underscan" \
-		        --backtitle "$SCRIPT_TITLE" \
-		        --yesno "Do you want to disable underscan (removes black borders)? Do this only if you're sure you need it!" 7 $DIALOG_WIDTH
-
-	        RESPONSE=$?
-	        case $RESPONSE in
-	           0) disbaleAtiUnderscan;;
-	           1) enableAtiUnderscan;;
-	           255) showInfo "ATI underscan configuration skipped";;
-	        esac
-        fi
         
-        showInfo "$VIDEO_MANUFACTURER_NAME video drivers successfully installed"
+        if [ $? == 0 ];
+        then
+            showInfo "$VIDEO_MANUFACTURER_NAME video drivers successfully installed"
+            
+            if [ $VIDEO_MANUFACTURER == "ati" ];
+            then
+	            sudo aticonfig --initial -f > /dev/null 2>&1
+	            sudo aticonfig --sync-vsync=on > /dev/null 2>&1
+	            sudo aticonfig --set-pcs-u32=MCIL,HWUVD_H264Level51Support,1 > /dev/null 2>&1
+
+	            dialog --title "Disable underscan" \
+		            --backtitle "$SCRIPT_TITLE" \
+		            --yesno "Do you want to disable underscan (removes black borders)? Do this only if you're sure you need it!" 7 $DIALOG_WIDTH
+
+	            RESPONSE=$?
+	            case $RESPONSE in
+                    0) 
+                        disbaleAtiUnderscan
+                        ;;
+                    1) 
+                        enableAtiUnderscan
+                        ;;
+                    255) 
+                        showInfo "ATI underscan configuration skipped"
+                        ;;
+	            esac
+            fi
+        else
+            showError "ERROR: $VIDEO_MANUFACTURER_NAME video drivers could not be installed"
+        fi
     else
 	    showInfo "Skipping. $VIDEO_MANUFACTURER_NAME video drivers already installed."
     fi
@@ -317,14 +431,26 @@ function disbaleAtiUnderscan()
 {
 	sudo kill $(pidof X) > /dev/null 2>&1
 	sudo aticonfig --set-pcs-val=MCIL,DigitalHDTVDefaultUnderscan,0 > /dev/null 2>&1
-    showInfo "Underscan successfully disabled"
+	
+	if [ $? == 0 ];
+    then
+        showInfo "Underscan successfully disabled"
+    else
+        showError "ERROR: Underscan could not be disabled"
+    fi
 }
 
 function enableAtiUnderscan()
 {
 	sudo kill $(pidof X) > /dev/null 2>&1
 	sudo aticonfig --set-pcs-val=MCIL,DigitalHDTVDefaultUnderscan,1 > /dev/null 2>&1
-    showInfo "Underscan successfully enabled"
+	
+	if [ $? == 0 ];
+    then
+        showInfo "Underscan successfully enabled"
+    else
+        showError "ERROR: Underscan could not be enabled"
+    fi
 }
 
 function installXbmcAutorunScript()
@@ -347,7 +473,13 @@ function installXbmcAutorunScript()
 	    sudo mv ./xbmc_init_script $INIT_FILE > /dev/null 2>&1
 	    sudo chmod a+x /etc/init.d/xbmc > /dev/null 2>&1
 	    sudo update-rc.d xbmc defaults > /dev/null 2>&1
-        showInfo "XBMC autorun succesfully configured"
+	    
+	    if [ $? == 0 ];
+        then
+            showInfo "XBMC autorun succesfully configured"
+        else
+            showError "ERROR: XBMC autorun could not be configured"
+        fi
 	fi
 }
 
@@ -359,27 +491,47 @@ function installXbmcBootScreen()
     if [ $? == 1 ];
     then
 	    sudo apt-get -y -qq install plymouth-label v86d > /dev/null 2>&1
-
-        mkdir -p $TEMP_DIRECTORY > /dev/null
-        cd $TEMP_DIRECTORY
-        wget -q $SCRIPTS_DIR_URL"plymouth-theme-xbmc-logo.deb" > /dev/null 2>&1
-        
-        if [ ! -f ./plymouth-theme-xbmc-logo.deb ];
+	    
+	    if [ $? == 0 ];
         then
-            showError "Download of XBMC boot screen package failed"
-        else
-            sudo dpkg -i ./plymouth-theme-xbmc-logo.deb > /dev/null 2>&1
-
-            if [ -f $INITRAMFS_SPLASH_FILE ];
+            mkdir -p $TEMP_DIRECTORY > /dev/null
+            cd $TEMP_DIRECTORY
+            wget -q $SCRIPTS_DIR_URL"plymouth-theme-xbmc-logo.deb" > /dev/null 2>&1
+            
+            if [ ! -f ./plymouth-theme-xbmc-logo.deb ];
             then
-                sudo rm $INITRAMFS_SPLASH_FILE > /dev/null 2>&1
-            fi
+                showError "Download of XBMC boot screen package failed"
+            else
+                sudo dpkg -i ./plymouth-theme-xbmc-logo.deb > /dev/null 2>&1
+                
+                if [ $? == 0 ];
+                then
+                    showInfo "XBMC boot screen successfully installed"
+                    
+                    if [ -f $INITRAMFS_SPLASH_FILE ];
+                    then
+                        sudo rm $INITRAMFS_SPLASH_FILE > /dev/null 2>&1
+                    fi
 
-            sudo touch $INITRAMFS_SPLASH_FILE > /dev/null 2>&1
-            echo "FRAMEBUFFER=y" | sudo tee -a $INITRAMFS_SPLASH_FILE > /dev/null 2>&1
-            sudo update-grub > /dev/null 2>&1
-            sudo update-initramfs -u > /dev/null 2>&1
-            showInfo "XBMC boot screen successfully installed"
+                    sudo touch $INITRAMFS_SPLASH_FILE > /dev/null 2>&1
+                    echo "FRAMEBUFFER=y" | sudo tee -a $INITRAMFS_SPLASH_FILE > /dev/null 2>&1
+                    sudo update-grub > /dev/null 2>&1
+                    sudo update-initramfs -u > /dev/null 2>&1
+                    
+                    if [ $? == 0 ];
+                    then
+                        showInfo "XBMC boot screen succesfully applied"
+                    else
+                        showError "ERROR: XBMC boot screen could not be applied"
+                    fi
+                else
+                    showError "ERROR: XBMC boot screen could not be installed"
+                fi
+            fi
+        
+            showInfo "XBMC boot screen installation requirements installed"
+        else
+            showError "ERROR: XBMC boot screen installation requirements could not be installed"
         fi
     else
         showInfo "Skipping. XBMC boot screen already installed"
@@ -407,7 +559,13 @@ function reconfigureXServer()
 
 	sudo touch $XWRAPPER_FILE > /dev/null 2>&1
 	echo "allowed_users=anybody" | sudo tee -a $XWRAPPER_CONFIG_FILE > /dev/null 2>&1
-	showInfo "X-server successfully configured"
+	
+	if [ $? == 0 ];
+    then
+        showInfo "X-server successfully configured"
+    else
+        showError "ERROR: X-server could not be configured"
+    fi
 }
 
 function selectAdditionalOptions()
